#607 refactoring matches path
